Let `indent-line-to` decide on (point) location
This unbreaks (newline-and-indent).

Before this commit calling newline-and-indent would put caret at the
beginning of line. This is not what a user wants, because they usually
desire continue typing code on the new line.

So remove the save-excursion call.

While at it, remove the beginning-of-line call as well, because both
are called first thing inside (pcal-mode--indent-column), so as far as
that function concerned having them earlier is unnecessary.