tla-mode: Let `indent-line-to` decide on (point) location
This is similar to

     fc1af6d250 "Let `indent-line-to` decide on (point) location"

It turns out there's two indenation functions, so fix the 2nd one as well.

This unbreaks (newline-and-indent).

Before this commit calling newline-and-indent would put caret at the
beginning of line. This is not what a user wants, because they usually
desire continue typing code on the new line.

So remove the save-excursion call.

While at it, remove the beginning-of-line call as well, because both
are called first thing inside (pcal-mode--indent-column), so as far as
that function concerned having them earlier is unnecessary.